# swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "RevUp Job Portal API",
                 "version":  "3.0.0",
                 "description":  "## RevUp — AI-Powered Job Portal API",
                 "contact":  {
                                 "name":  "RevUp Dev Team"
                             }
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5000",
                        "description":  "Local Development Server"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT",
                                                                  "description":  "Enter your JWT token. Obtain it from POST /api/auth/login"
                                                              }
                                           },
...
                   },
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/applications/my-applications":  {
                                                            "get":  {
                                                                        "tags":  [
                                                                                     "Applications"
                                                                                 ],
                                                                        "summary":  "Get seeker\u0027s own application history",
                                                                        "security":  [
                                                                                         {
                                                                                             "bearerAuth":  [

                                                                                                            ]
                                                                                         }
                                                                                     ],
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "List of applications with job details"
                                                                                                  }
                                                                                      }
                                                                    }
                                                        },
                  "/api/applications/job/{jobId}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Applications"
                                                                             ],
                                                                    "summary":  "Get all applicants for a specific job (Recruiter only)",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "in":  "path",
                                                                                           "name":  "jobId",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "integer"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Applicants list"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "description":  "Forbidden - Not your job"
                                                                                              }
                                                                                  }
                                                                }
                                                    },
                  "/api/applications/apply/{jobId}":  {
                                                          "post":  {
                                                                       "tags":  [
                                                                                    "Applications"
                                                                                ],
                                                                       "summary":  "Apply to a job (with optional resume upload)",
                                                                       "security":  [
                                                                                        {
                                                                                            "bearerAuth":  [

                                                                                                           ]
                                                                                        }
                                                                                    ],
                                                                       "parameters":  [
                                                                                          {
                                                                                              "in":  "path",
                                                                                              "name":  "jobId",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "requestBody":  {
                                                                                           "content":  {
                                                                                                           "multipart/form-data":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "type":  "object",
                                                                                                                                                      "properties":  {
                                                                                                                                                                         "cover_letter":  {
                                                                                                                                                                                              "type":  "string"
                                                                                                                                                                                          },
                                                                                                                                                                         "resume":  {
                                                                                                                                                                                        "type":  "string",
                                                                                                                                                                                        "format":  "binary"
                                                                                                                                                                                    }
                                                                                                                                                                     }
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                       "responses":  {
                                                                                         "201":  {
                                                                                                     "description":  "Application submitted"
                                                                                                 },
                                                                                         "409":  {
                                                                                                     "description":  "Already applied"
                                                                                                 }
                                                                                     }
                                                                   }
                                                      },
                  "/api/applications/{id}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "Applications"
                                                                      ],
                                                             "summary":  "View specific application with seeker profile (Recruiter — Privacy Rule enforced)",
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ],
                                                             "parameters":  [
                                                                                {
                                                                                    "in":  "path",
                                                                                    "name":  "id",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "integer"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "Application detail with seeker info"
                                                                                       },
                                                                               "403":  {
                                                                                           "description":  "Forbidden"
                                                                                       }
                                                                           }
                                                         },
                                                 "delete":  {
                                                                "tags":  [
                                                                             "Applications"
                                                                         ],
                                                                "summary":  "Seeker withdraws their application",
                                                                "security":  [
                                                                                 {
                                                                                     "bearerAuth":  [

                                                                                                    ]
                                                                                 }
                                                                             ],
                                                                "parameters":  [
                                                                                   {
                                                                                       "in":  "path",
                                                                                       "name":  "id",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "integer"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Application withdrawn"
                                                                                          },
                                                                                  "400":  {
                                                                                              "description":  "Can only withdraw applied status"
                                                                                          },
                                                                                  "403":  {
                                                                                              "description":  "Forbidden - Not your application"
                                                                                          }
                                                                              }
                                                            }
                                             },
                  "/api/applications/{id}/status":  {
                                                        "put":  {
                                                                    "tags":  [
                                                                                 "Applications"
                                                                             ],
                                                                    "summary":  "Update application status and trigger notification + email (Recruiter only)",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "in":  "path",
                                                                                           "name":  "id",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "integer"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "requestBody":  {
                                                                                        "required":  true,
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "object",
                                                                                                                                                "required":  [
                                                                                                                                                                 "status"
                                                                                                                                                             ],
                                                                                                                                                "properties":  {
                                                                                                                                                                   "status":  {
                                                                                                                                                                                  "type":  "string",
                                                                                                                                                                                  "enum":  [
                                                                                                                                                                                               "applied",
                                                                                                                                                                                               "shortlisted",
                                                                                                                                                                                               "rejected",
                                                                                                                                                                                               "hired"
                                                                                                                                                                                           ]
                                                                                                                                                                              }
                                                                                                                                                               }
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Status updated, notification sent"
                                                                                              }
                                                                                  }
                                                                }
                                                    },
...
              }
}